Refatorar paginas client para evitar exportacao de metadata
