+++ app/login/page.tsx
@@ -1,12 +1,6 @@
-"use client";
-
-import Image from "next/image";
-import styles from "../../styles/Home.module.css";
-import logo from "../../public/logo.png";
-import { useEffect } from "react";
-
 import type { Metadata } from "next";
 import { defaultMetadata } from "@/lib/seo";
+import LoginClient from "./LoginClient";
 
 export const metadata: Metadata = {
   ...defaultMetadata,
@@ -15,79 +9,6 @@ export const metadata: Metadata = {
     "Conecte sua conta do Patreon para acessar recursos premium do servidor UO Babel.",
 };
 
-import { useRouter } from "next/navigation";
-export default function LoginPage() {
-  const clientId = process.env.NEXT_PUBLIC_PATREON_CLIENT_ID;
-  const rawRedirectUri = "https://www.uobabel.com/patreon/callback";
-  const encodedRedirectUri = encodeURIComponent(rawRedirectUri);
-  const scope = encodeURIComponent("identity identity.memberships");
-  const router = useRouter();
-
-  useEffect(() => {
-    async function verificarStatus() {
-      try {
-        const res = await fetch("/api/patreon/me");
-        const data = await res.json();
-
-        if (data.success && data.user?.isSubscriber) {
-          // Redireciona automaticamente se já for assinante
-          router.push("/painel");
-        }
-      } catch (err) {
-        console.warn("Falha ao verificar status do Patreon:", err);
-      }
-    }
-
-    verificarStatus();
-  }, [router]);
-  // ⛔ Aviso se clientId não estiver definido
-  if (!clientId) {
-    return (
-      <div className={styles.container}>
-        <p style={{ color: "red", textAlign: "center", marginTop: "2rem" }}>
-          ⚠️ Erro: NEXT_PUBLIC_PATREON_CLIENT_ID não definido. Verifique seu .env.local ou
-          variáveis no Vercel.
-        </p>
-      </div>
-    );
-  }
-
-  // 🔗 URL segura de login com OAuth
-  const loginUrl = `https://www.patreon.com/oauth2/authorize?response_type=code&client_id=${clientId}&redirect_uri=${encodedRedirectUri}&scope=${scope}&prompt=consent`;
-
-  return (
-    <div className={styles.container}>
-
-      <header className={styles.hero}>
-        <div className={styles.heroContent}>
-          <Image
-            src={logo}
-            alt="UO Babel Logo"
-            className={styles.logo}
-            width={180}
-            height={180}
-          />
-          <h1 className={styles.title}>UO Babel</h1>
-          <p className={styles.subtitle}>Servidor Survival Hardcore</p>
-        </div>
-      </header>
-
-      <main className={styles.main}>
-        <section className={styles.intro}>
-          <h2>🔐 Acesso Premium</h2>
-          <p>
-            Para acessar recursos exclusivos no jogo, entre com sua conta do
-            Patreon.
-          </p>
-          <a href={loginUrl} className={styles.btn}>
-            Entrar com Patreon
-          </a>
-        </section>
-      </main>
-
-      <footer className={styles.footer}>
-        <p>&copy; 2025 UO Babel. Todos os direitos reservados.</p>
-      </footer>
-    </div>
-  );
+export default function Page() {
+  return <LoginClient />;
 }
